telephony: Support for KSC5601 on SIM.
cherry-picked from f1c01bd69e6d209e15ddbd90f1fd2265715a7e57

Korean phones write to the ADN record of the SIM in a non-standard way.
When UCS2 is not used, the alphaTag will be written in the KSC5601
encoding. This contribution adds support for reading that format when
a Korean SIM card is used.

Also adds support for KSC5601 in SMS.

Change-Id: I81a4a6949359b4d23a937ac2d813bafed2b85ff6

Conflicts:

	telephony/java/com/android/internal/telephony/IccUtils.java
	telephony/java/com/android/internal/telephony/gsm/SmsMessage.java